Projects-001-BE/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI & Vector
pgvector>=0.3.0
google-genai>=1.70.0
google-auth>=2.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.18
httpx>=0.28.0
